Revamping automata package: first iteration
