--- automata/README.md
@@ -1,7 +1,0 @@
-# Automata
-
-Finite automata are recognizers; they simply say yes or no for each possible input string.
-They come in two flavors:
-
-  - Deterministic finite automata (DFA)
-  - Non-deterministic finite automata (NFA)
