+++ automata/automata.go
@@ -20,7 +20,7 @@ var (
 	CmpSymbol  = generic.NewCompareFunc[Symbol]()
 	HashSymbol = hash.HashFuncForInt32[Symbol](nil)
 
-	eqStateSet = func(a, b States) bool {
+	eqStates = func(a, b States) bool {
 		return a.Equal(b)
 	}
 
@@ -61,67 +61,8 @@ func NewSymbols(a ...Symbol) set.Set[Symbol] {
 // String represents a sequence of symbols in an automaton.
 type String []Symbol
 
-// toString creates a string of symbols from a string.
-func toString(s string) String {
-	res := make(String, len(s))
-	for i, r := range s {
-		res[i] = Symbol(r)
-	}
-
-	return res
-}
-
-// Transition represents a transition in a finite automaton.
-// It can be either a DFA transition with a single next state or an NFA transition with multiple next states.
-type Transition[T State | []State] struct {
-	State
-	Symbol
-	Next T
-}
-
-// stateManager is used for keeping track of states when combining multiple automata.
-type stateManager struct {
-	last   State
-	states map[int]map[State]State
-}
-
-func newStateManager(last State) *stateManager {
-	return &stateManager{
-		last:   last,
-		states: map[int]map[State]State{},
-	}
-}
-
-func (m *stateManager) GetOrCreateState(id int, s State) State {
-	if _, ok := m.states[id]; !ok {
-		m.states[id] = make(map[State]State)
-	}
-
-	if t, ok := m.states[id][s]; ok {
-		return t
-	}
-
-	m.last++
-	m.states[id][s] = m.last
-	return m.last
-}
-
-// generatePermutations generates all permutations of a sequence of states using recursion and backtracking.
-// Each permutation is passed to the provided yield function.
-func generatePermutations(states []State, start, end int, yield func([]State) bool) bool {
-	if start == end {
-		return yield(states)
-	}
-
-	for i := start; i <= end; i++ {
-		states[start], states[i] = states[i], states[start]
-		cont := generatePermutations(states, start+1, end, yield)
-		states[start], states[i] = states[i], states[start]
-
-		if !cont {
-			return false
-		}
-	}
-
-	return true
+// SymbolRange represents a range of symbols inclusive.
+type SymbolRange struct {
+	Start Symbol
+	End   Symbol
 }
