+++ automata/util_test.go
@@ -1,0 +1,106 @@
+package automata
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+// toString creates a string of symbols from a string.
+// It is a helper function for creating test cases.
+func toString(s string) String {
+	res := make(String, len(s))
+	for i, r := range s {
+		res[i] = Symbol(r)
+	}
+
+	return res
+}
+
+func TestGenerateStatePermutations(t *testing.T) {
+	tests := []struct {
+		name                 string
+		states               []State
+		start                int
+		end                  int
+		result               bool
+		expectedResult       bool
+		expectedPermutations [][]State
+	}{
+		{
+			name:           "OK",
+			states:         []State{0, 1, 2},
+			start:          0,
+			end:            2,
+			result:         true,
+			expectedResult: true,
+			expectedPermutations: [][]State{
+				{0, 1, 2},
+				{0, 2, 1},
+				{1, 0, 2},
+				{1, 2, 0},
+				{2, 1, 0},
+				{2, 0, 1},
+			},
+		},
+		{
+			name:           "ReturnEarly",
+			states:         []State{0, 1, 2},
+			start:          0,
+			end:            2,
+			result:         false,
+			expectedResult: false,
+			expectedPermutations: [][]State{
+				{0, 1, 2},
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			perms := [][]State{}
+			result := generateStatePermutations(tc.states, tc.start, tc.end, func(perm []State) bool {
+				clone := make([]State, len(perm))
+				copy(clone, perm)
+
+				perms = append(perms, clone)
+				return tc.result
+			})
+
+			assert.Equal(t, tc.expectedResult, result)
+			assert.Equal(t, tc.expectedPermutations, perms)
+		})
+	}
+}
+
+func TestStateManager(t *testing.T) {
+	tests := []struct {
+		name          string
+		last          State
+		id            int
+		s             State
+		expectedState State
+	}{
+		{
+			name:          "OK",
+			last:          10,
+			id:            0,
+			s:             1,
+			expectedState: 11,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			sm := newStateManager(tc.last)
+
+			// Test creating a new state
+			state := sm.GetOrCreateState(tc.id, tc.s)
+			assert.Equal(t, tc.expectedState, state)
+
+			// Test retrieving the same state again
+			state = sm.GetOrCreateState(tc.id, tc.s)
+			assert.Equal(t, tc.expectedState, state)
+		})
+	}
+}
