+++ graph/README.md
@@ -1,3 +1,6 @@
+# Graph
+
+## Test Cases
 
 The example graphs used in test files are shown below.
